refactor(core): split render-svg into focused modules
Extract the 530-line pure SVG renderer into format helpers, mark emitters,
scene/chrome assembly, and a thin public barrel. Capture a pre-split kitchen-
sink golden fixture so vertical extracts cannot drift markup, class names, or
ids. Public API (countMarks, pathData, renderToSVGString, sceneLabel,
sceneToSVGString, RenderSVGOptions) stays on the package root via the barrel.
